ADD feature : separate C++ .h and .cpp
ADD feature : separate C++ .h and .cpp without logic change.
